suggestion: Refactor
- Fix a typo in hyperband service `MaxParallel`
- Move the service to the suggestion package
- Move the port to the suggestion package
