+++ suggestion/hyperband/main.go
@@ -1,27 +1,26 @@
 package main
 
 import (
-	pb "github.com/kubeflow/hp-tuning/api"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/reflection"
 	"log"
 	"net"
-)
 
-const (
-	port = "0.0.0.0:6789"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/reflection"
+
+	pb "github.com/kubeflow/hp-tuning/api"
+	"github.com/kubeflow/hp-tuning/suggestion"
 )
 
 func main() {
-	listener, err := net.Listen("tcp", port)
+	listener, err := net.Listen("tcp", suggestion.DefaultPort)
 	if err != nil {
 		log.Fatalf("Failed to listen: %v", err)
 	}
 	size := 1<<31 - 1
 	s := grpc.NewServer(grpc.MaxRecvMsgSize(size), grpc.MaxSendMsgSize(size))
-	pb.RegisterSuggestionServer(s, NewHyperBandSuggestService())
+	pb.RegisterSuggestionServer(s, suggestion.NewHyperBandSuggestService())
 	reflection.Register(s)
-	log.Printf("HyperBand Suggestion Service\n")
+	log.Println("HyperBand Suggestion Service")
 	if err = s.Serve(listener); err != nil {
 		log.Fatalf("Failed to serve: %v", err)
 	}
